requirements.txt
asgiref==3.6.0
attrs==23.1.0
azure-common==1.1.28
azure-core==1.29.2
azure-mgmt-core==1.4.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==3.2.0
distlib==0.3.6
Django==4.1.5
djangorestframework==3.14.0
exceptiongroup==1.1.3
filelock==3.8.0
h11==0.14.0
idna==3.4
isodate==0.6.1
outcome==1.2.0
pipenv==2022.11.23
platformdirs==2.5.4
pycparser==2.21
PySocks==1.7.1
pytz==2023.3
requests==2.31.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.2
trio-websocket==0.10.3
typing-extensions==4.7.1
tzdata==2022.7
urllib3==2.0.4
virtualenv==20.16.7
virtualenv-clone==0.5.7
wsproto==1.2.0
